requirements.txt
fastapi==0.115.6
langchain==0.3.10
langchain-cohere==0.3.3
langchain-community==0.3.10
pandas==2.2.3
python-dotenv==1.0.1
ruff==0.8.2
uvicorn==0.32.1